refactor categories' form
