+++ src/app/screens/Categories/EditCategory/EditCategory.tsx
@@ -1,45 +1,32 @@
-import { Link, useNavigate, useParams } from "react-router-dom";
+import { /* Link, */ useNavigate, useParams } from "react-router-dom";
 import styles from "./EditCategory.module.scss";
 import { useTheme } from "../../../hooks/useTheme";
-import { useEffect, useState } from "react";
+/* import { useEffect, useState } from "react"; */
 import { CategorySchemaCreate } from "../../../interfaces/interfaces";
 import { API_CATEGORIES } from "../../../constants/urlsAPI";
-import { useMutation } from "react-query";
+import { useMutation, useQuery } from "react-query";
+import { FormCategory } from "../../../components/Forms/FormCategory/FormCategory";
+import { QUERY_KEY_CATEGORIES } from "../../../constants/queryConstants";
+import { Loader } from "../../../components/UI/Loader/Loader";
+import { ErrorComponent } from "../../../components/Error/ErrorComponent";
 
 export function EditCategory() {
   const [darkMode] = useTheme();
   const { id } = useParams();
-  const [name, setName] = useState<string>("");
-  const [image, setImage] = useState<string>("");
   const navigate = useNavigate();
 
-  const [errorName, setErrorName] = useState<boolean>();
-
-  useEffect(() => {
-    getUniqueCategory.mutate(id);
-  }, []);
-  const getUniqueCategory = useMutation(
-    async (id: string | undefined) => {
-      const res = await fetch(API_CATEGORIES + `/${id}`, {
-        method: "GET",
-      });
-      const json = await res.json();
-      if (json.statusCode === 401) {
-        throw new Error("Category not found");
-      }
-      return json;
-    },
-    {
-      onSuccess: (data) => {
-        setName(data?.name);
-        setImage(data?.image);
-      },
-      onError: (error) => {
-        console.error("Error " + error);
-        navigate(`/categories`, { replace: true });
-      },
+  const {
+    data: category,
+    status: categoryStatus,
+    error: categoryError,
+  } = useQuery([QUERY_KEY_CATEGORIES, id], async () => {
+    const res = await fetch(API_CATEGORIES + `/${id}`);
+    const json = await res.json();
+    if (json.error == "Not Found") {
+      throw new Error(json);
     }
-  );
+    return json;
+  });
 
   const updateCategory = useMutation(
     async (data: CategorySchemaCreate) => {
@@ -64,27 +51,6 @@ export function EditCategory() {
     }
   );
 
-  function handleBlurName() {
-    if (name.length > 0) {
-      setErrorName(false);
-    } else {
-      setErrorName(true);
-    }
-  }
-  function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-    let urlImage = image;
-    if (urlImage.length == 0) {
-      urlImage = "https://placeimg.com/640/480/any?r=0.9178516507833767";
-    }
-    const category: CategorySchemaCreate = {
-      name: name,
-      image: urlImage,
-    };
-
-    updateCategory.mutate(category);
-  }
-
   return (
     <main
       className={[
@@ -92,54 +58,19 @@ export function EditCategory() {
         darkMode ? styles.darkMode : styles.lightMode,
       ].join(" ")}
     >
-      <form className={styles.form} onSubmit={handleSubmit}>
-        <label htmlFor="inputName">
-          name:{" "}
-          <input
-            type="text"
-            name="name"
-            id="inputName"
-            placeholder="name"
-            value={name}
-            onChange={(e) => {
-              setName(e.target.value);
-            }}
-            onBlur={handleBlurName}
-            aria-errormessage="nameErrorID"
-            aria-invalid={!errorName}
-            required
-          />
-          <small
-            id="msg-exist-name-ID"
-            aria-live="assertive"
-            style={{ visibility: errorName ? "visible" : "hidden" }}
-          >
-            The category must be contain a name.
-          </small>
-        </label>
-        <label htmlFor="inputImage">
-          image's URL:{" "}
-          <input
-            type="url"
-            name="image"
-            id="inputImage"
-            placeholder="URL of image"
-            value={image}
-            onChange={(e) => {
-              setImage(e.target.value);
-            }}
-          />
-        </label>
-        <button
-          disabled={errorName}
-          type="submit"
-          className={styles.btnRegister}
-        >
-          Confirm changes
-        </button>
-        {" or "}
-        <Link to="/categories">Cancel</Link>
-      </form>
+      {categoryError ? (
+        <ErrorComponent />
+      ) : (
+        <>
+          {categoryStatus === "loading" && <Loader />}
+          {category && (
+            <FormCategory
+              onSubmit={updateCategory}
+              category={category}
+            ></FormCategory>
+          )}
+        </>
+      )}
     </main>
   );
 }
